@@ -2,6 +2,11 @@
   "name": "alis",
   "version": "1.0.0",
   "license": "Unlicense",
+  "type": "module",
+  "engines": {
+    "node": "^20.10.0",
+    "pnpm": "^8.12.1"
+  },
   "scripts": {
     "dev": "next dev",
     "build": "prisma generate && next build",
@@ -16,28 +21,28 @@
     "schema": "src/database/schema.prisma"
   },
   "devDependencies": {
-    "@types/node": "^18.11.13",
-    "@types/nprogress": "^0.2.0",
-    "@types/react": "^18.0.26",
-    "cypress": "^12.0.2",
-    "eslint-config-silash35": "^1.2.1",
-    "prisma": "^4.7.1",
-    "sass": "^1.56.2",
-    "start-server-and-test": "^1.15.2",
-    "typescript": "^4.9.4"
+    "@types/node": "^20.11.0",
+    "@types/nprogress": "^0.2.3",
+    "@types/react": "^18.2.47",
+    "cypress": "^13.6.2",
+    "eslint-config-silash35": "^3.3.1",
+    "prisma": "^5.8.0",
+    "sass": "^1.69.7",
+    "start-server-and-test": "^2.0.3",
+    "typescript": "^5.3.3"
   },
   "dependencies": {
-    "@emotion/react": "^11.10.5",
-    "@emotion/styled": "^11.10.5",
-    "@mui/icons-material": "^5.10.16",
-    "@mui/material": "^5.10.17",
-    "@prisma/client": "^4.7.1",
-    "next": "^13.0.6",
-    "next-auth": "^4.18.6",
+    "@emotion/react": "^11.11.3",
+    "@emotion/styled": "^11.11.0",
+    "@mui/icons-material": "^5.15.4",
+    "@mui/material": "^5.15.4",
+    "@prisma/client": "^5.8.0",
+    "next": "^14.0.4",
+    "next-auth": "^4.24.5",
     "next-pwa": "^5.6.0",
     "nprogress": "^0.2.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "swr": "^2.0.0"
+    "swr": "^2.2.4"
   }
 }
