@@ -3,20 +3,20 @@ import Link from "next/link";
 import Account from "./Account";
 import styles from "./header.module.scss";
 
-export default function Header() {
-  return (
-    <header className={styles.header}>
-      <nav>
-        <div className={styles.links}>
-          <Link href="/" className={styles.logo}>
-            a
-          </Link>
-        </div>
+const Header = () => (
+  <header className={styles.header}>
+    <nav>
+      <div className={styles.links}>
+        <Link className={styles.logo} href="/">
+          a
+        </Link>
+      </div>
 
-        <div className={styles.links}>
-          <Account />
-        </div>
-      </nav>
-    </header>
-  );
-}
+      <div className={styles.links}>
+        <Account />
+      </div>
+    </nav>
+  </header>
+);
+
+export default Header;
