@@ -9,22 +9,22 @@ interface Props {
   place: Place;
 }
 
-export default function Card({ place }: Props) {
-  return (
-    <Link href={`/places/${place.id}`}>
-      <article
-        className={styles.card}
-        style={{
-          backgroundImage: `url(${place.imageURL}), url(/images/empty.png)`,
-        }}
-      >
-        <div className={styles.content}>
-          <h2>{place.name}</h2>
-          <IconButton aria-label="Discover the place">
-            <Explore fontSize="large" />
-          </IconButton>
-        </div>
-      </article>
-    </Link>
-  );
-}
+const Card = ({ place }: Props) => (
+  <Link href={`/places/${place.id}`}>
+    <article
+      style={{
+        backgroundImage: `url(${place.imageURL}), url(/images/empty.png)`,
+      }}
+      className={styles.card}
+    >
+      <div className={styles.content}>
+        <h2>{place.name}</h2>
+        <IconButton aria-label="Discover the place">
+          <Explore fontSize="large" />
+        </IconButton>
+      </div>
+    </article>
+  </Link>
+);
+
+export default Card;
