@@ -1,12 +1,10 @@
-const Love = () => {
-  return (
-    <svg fill="#222222" width="296" height="300" version="1.1" xmlns="http://www.w3.org/2000/svg">
-      <g transform="matrix(1.4213 0 0 1.4213 -36.37 -21.406)">
-        <path d="m74.904 221.56c-0.18582-0.19116 0.02552-3.3476 0.46975-7.0142 1.2566-10.373 0.40827-23.378-1.8165-27.845-4.1258-8.2837-5.6136-15.201-7.847-36.48-2.5531-24.326-2.5558-24.337-12.097-51.954-7.2975-21.123-7.5036-22.275-5.1268-28.65 0.94139-2.525 4.0607-4.4793 7.1493-4.4793 4.6304 0 6.8049 2.7765 12.903 16.475 3.2707 7.3471 7.3576 16.456 9.082 20.243l3.1353 6.8846 0.76541-3.5378c1.7508-8.0921 7.0963-12.173 14.735-11.25 5.1089 0.61774 4.5571 0.93618 9.9236-5.7272 1.981-2.4598 3.3054-3.1698 6.6315-3.5555 2.2859-0.26508 5.121-0.23283 6.3002 0.07169 3.2842 0.84801 6.536 4.751 7.7864 9.3458 0.62101 2.2819 1.3748 3.8965 1.6752 3.5879 0.30034-0.30853 0.74682-15.914 0.9922-34.679 0.42267-32.324 0.54259-34.287 2.2806-37.331 0 0 0.91844-2.3872 1.8345-3.2128 1.8115-1.6326 4.2563-2.957 6.6949-2.9498 2.4064 0.0071 4.863 1.273 6.5889 2.9498 1.1042 1.0728 2.008 4.1591 2.008 4.1591 1.9523 4.0436 2.0332 5.3564 2.9132 47.274 0.49787 23.713 1.1164 44.561 1.3745 46.328 0.87291 5.9757 2.4478 10.061 4.0847 10.595 1.7786 0.58074 16.207-6.6556 15.528-7.7873-0.22012-0.36643 0.0919-0.66623 0.69333-0.66623 0.60145 0 2.7289-1.1134 4.7276-2.4741 5.307-3.6131 12.48-6.11 19.084-6.6424 4.9645-0.40027 6.3136-0.15897 9.3687 1.6756 3.8581 2.3169 6.2172 6.1713 5.3994 8.8221-0.28127 0.91174-2.2135 2.6166-4.2938 3.7886s-7.4322 5.9391-11.893 10.594c-4.9322 5.1462-10.527 9.919-14.276 12.179-7.7171 4.6513-15.614 11.984-23.861 22.155-7.6822 9.4756-9.4431 11.102-15.615 14.428-2.617 1.4099-5.2995 3.6404-5.9611 4.9566-0.73759 1.4674-1.448 7.5788-1.8364 15.798-0.55866 11.822-0.8233 13.456-2.2406 13.838-0.88394 0.2378-1.8312 0.20192-2.105-0.0798-0.27378-0.28166-0.49779-6.7894-0.49779-14.462v-13.95l-5.5766-0.484c-3.0671-0.26621-7.3834-1.1792-9.5918-2.029-2.2083-0.84975-4.5171-1.5835-5.1305-1.6306-0.61342-0.0471-0.93163-0.27454-0.70713-0.50551 0.2245-0.23096-0.47815-1.1592-1.5614-2.0628-3.4436-2.8724-0.94536-3.0045 5.7865-0.30608 10.577 4.2399 13.529 3.8283 25.343-3.5336 9.9919-6.2265 16.812-12.568 17.779-16.531 0.28922-1.1855 0.10206-3.3022-0.41589-4.7037-1.1771-3.1851-1.2308-8.8544-0.0839-8.8544 0.47182 0 1.3155 1.7556 1.8749 3.9013 0.55939 2.1457 1.2006 4.2108 1.425 4.589 0.22436 0.37825 3.2954-1.818 6.8246-4.8806 3.5292-3.0625 9.386-7.427 13.015-9.6988 4.1185-2.5782 9.4639-7.2126 14.223-12.331 4.1934-4.5102 8.3065-8.2004 9.1403-8.2004 0.83376 0 2.388-0.7265 3.4539-1.6144 2.4287-2.0233 1.7258-4.2521-1.8172-5.7619-4.7143-2.009-18.83 2.0141-24.791 7.0658-2.0105 1.7036-2.0385 1.8723-0.56297 3.3903 1.9257 1.9811 1.2526 2.8277-1.4923 1.8769-2.8372-0.9828-4.823-0.58034-10.551 2.1385-2.6991 1.2811-7.0154 3.3016-9.5918 4.4901s-4.7346 2.5791-4.7959 3.0904c-0.0613 0.51124-0.21191 1.3426-0.3346 1.8475-0.12268 0.50487-0.27325 1.3434-0.33459 1.8634-0.0613 0.51998-1.2922 0.0474-2.7352-1.0502-2.1881-1.6643-3.3362-1.8961-6.915-1.396-7.488 1.0463-12.259 3.121-13.95 6.0661-1.6963 2.9542-6.6386 4.9336-10.092 4.0418-3.004-0.77568-8.0112 1.2638-10.463 4.2617-1.1916 1.4569-2.6859 4.7339-3.3208 7.2823-0.65971 2.6482-1.6627 4.6334-2.341 4.6334-1.2146 0-1.1577-0.63006 0.77309-8.5575 0.55247-2.2684 2.3121-4.9456 4.9776-7.5731 3.94-3.8838 4.3859-5.5004 0.73336-2.6581-2.5494 1.9838-8.7309 1.7116-11.47-0.50511-1.5512-1.2553-3.6447-5.4739-6.4916-13.081-7.8955-21.097-7.5725-20.667-13.229-17.631-3.149 1.6895-4.4491 0.78245-1.6518-1.1523 2.6242-1.8151 2.5351-3.2273-0.63978-10.132-2.7335-5.9444-4.9171-7.8522-6.4374-5.6245-0.59928 0.8781-0.65932 0.41476-0.19086-1.4726 1.2275-4.9455 0.72312-8.0353-1.6389-10.04-1.4752-1.2519-2.4303-3.2411-2.8187-5.8706-0.68261-4.6211-4.5153-12.87-6.309-13.578-2.145-0.8468-4.8438 1.3985-5.391 4.4852-0.53974 3.0442 1.8795 10.771 10.271 32.805 5.1947 13.64 5.1496 13.535 6.0736 14.123 0.47647 0.30295 0.63883 4.0928 0.37741 8.8097-0.25233 4.5529-0.30271 8.2774-0.11195 8.2767 0.19075-6.9e-4 3.7322-0.61418 7.8699-1.3633 4.1377-0.74913 7.7179-1.1616 7.956-0.91669 0.60964 0.62719 0.39288 0.75173-6.352 3.6493-7.2396 3.1102-7.3755 3.453-6.5533 16.536 1.0063 16.012 4.4325 25.334 12.144 33.042l3.6739 3.6718-5.008 0.93068-1.4858 12.845c-0.8172 7.0647-1.5199 13.107-1.5615 13.426-0.08887 0.68308-3.3854 1.6742-3.8792 1.1662zm49.714-74.678c1.8294-2.0796 1.9233-8.1504 0.48358-31.257-1.0845-17.406-2.1058-22.356-5.1148-24.791-4.7953-3.8806-11.5-1.0648-14.092 5.9184-1.5534 4.1848-1.54 4.3949 1.2224 19.125 1.5362 8.1916 3.2823 17.785 3.8801 21.319 0.81703 4.83 1.6896 7.0525 3.5137 8.95 2.9747 3.0943 7.7278 3.44 10.107 0.73499zm-2.1897-4.8562c0-4.2063-0.0388-4.2625-3.1415-4.5438-3.8751-0.35143-4.3798-1.6834-1.0681-2.8191 1.8822-0.64548 3.0401-0.47523 4.7898 0.70421 2.0211 1.3624 2.2478 2.0136 1.8263 5.2462-0.73388 5.629-2.4065 6.6108-2.4065 1.4125zm-6.2458-22.742c1.4738-0.96498 6.8848-1.2558 5.9764-0.32123-0.31178 0.32075-2.0851 0.72141-3.9408 0.89035-2.3987 0.21839-2.9871 0.0539-2.0356-0.56912zm-8.332 25.159c1.3572-2.1309 1.4088-3.2109 0.4719-9.8679-1.5304-10.874-7.0715-32.764-8.7875-34.715-3.0438-3.4602-10.775-1.1886-12.764 3.7501-2.158 5.3598-1.2624 11.815 3.5849 25.839 5.9881 17.325 6.0288 17.386 11.55 17.386 3.7785 0 4.6414-0.34736 5.9441-2.3928zm-2.4132-4.4918c-1.171-2.361-1.8434-2.7384-4.7139-2.6459-3.692 0.11897-4.0063-1.2633-0.67871-2.985 4.0517-2.0963 7.6152 0.89231 7.0419 5.9058l-0.28346 2.479zm-6.8598-19.454c0.86814-0.23273 2.0727-0.21743 2.6768 0.034 0.60409 0.25145-0.10621 0.44187-1.5784 0.42316-1.4722-0.0187-1.9665-0.22444-1.0983-0.45717zm43.258 17.591c4.5394-0.86772 8.2534-1.8888 8.2534-2.269s-2.8106-1.4413-6.2458-2.3579c-3.4352-0.91664-6.2458-2.0353-6.2458-2.4859 0-0.46392 2.8616-0.54344 6.5971-0.18331l6.5971 0.636-1.3621-5.3769c-0.94743-3.7401-1.5297-12.333-1.9127-28.228-0.4882-20.259-0.72521-22.964-2.0894-23.847-1.1131-0.72072-2.2856-0.70636-4.2382 0.05192-2.0992 0.81516-2.6995 0.79565-2.6995-0.08775 0-0.62481 1.9072-1.7441 4.2382-2.4873l4.2382-1.3513 0.26455-7.2978c0.33429-9.2211-0.1686-11.26-2.9549-11.979-2.3486-0.60642-3.0653-2.2093-0.98792-2.2093 2.9706 0 3.4699-2.4079 2.4124-11.633-0.56124-4.8958-1.4586-9.4447-1.9941-10.108-1.7035-2.1117-5.9075-1.4582-7.5874 1.1793-1.3226 2.0767-1.6175 6.7242-2.273 35.818-0.5037 22.357-0.45287 34.009 0.15343 35.175 0.49866 0.95858 2.1099 2.0148 3.5806 2.3471 4.8397 1.0936 3.8459 2.2975-1.9373 2.347-5.378 0.046-5.5396 0.11245-3.6217 1.4889 2.8845 2.0701 2.5036 3.3106-0.78812 2.5668l-2.7957-0.63171 0.60332 14.098c0.73785 17.243 1.0815 19.356 3.0591 18.811 0.81556-0.22456 5.1969-1.1182 9.7362-1.986zm-49.96 59.654c5.4706-3.9421 11.543-5.0146 17.259-3.0484 5.5026 1.8927 4.7563 2.5594-2.845 2.5417-4.1282-0.01-8.9085 0.41905-10.623 0.95261-4.5946 1.43-6.1421 1.2479-3.7914-0.44595z" />
-        <path d="m78.915 31.501c-1.8054 0.50275-3.5949 1.6255-4.9013 3.0777-0.72939 0.81231-1.5317 2.0888-1.977 3.1462-0.98597 2.3319-1.3168 4.8578-1.0266 7.8386 0.50371 5.1633 3.0812 8.4592 8.7104 11.14 1.0386 0.49476 1.9904 0.90056 3.9908 1.6965 2.754 1.095 3.8264 1.5753 5.2735 2.3578 3.2651 1.7602 6.57 4.3695 7.9739 6.2921 0.16904 0.23242 0.34178 0.48663 0.38617 0.56455l0.08226 0.14344 0.04324-0.231c0.32739-1.8311 2.7241-5.6679 5.5787-8.9428 1.2507-1.4342 2.3739-2.5163 4.8364-4.6572 2.9368-2.5536 4.1044-3.7235 5.289-5.3127 1.7923-2.3997 2.5647-4.6567 2.5434-7.4302-5e-3 -0.69338-0.36744-2.9011-0.61764-3.7654-0.61029-2.0959-1.3487-3.6515-2.4581-5.156-0.78363-1.0601-1.6757-1.9279-2.7456-2.6684-1.8797-1.2992-3.8903-1.9098-6.2222-1.8907-0.59946 0.0075-2.3352 0.30186-2.9222 0.4954-3.3687 1.132-5.8962 3.5657-7.4543 7.1726-0.21002 0.48189-0.56352 1.4931-0.70263 2.0218l-0.11004 0.40301-0.58821-0.80506c-0.32142-0.44514-0.78883-1.0368-1.0383-1.3189-2.7542-3.106-6.2141-4.7048-9.8106-4.5212-0.47056 0.02246-1.6709 0.2212-2.1331 0.34984z" />
-      </g>
-    </svg>
-  );
-};
+const Love = () => (
+  <svg fill="#222222" height="300" version="1.1" width="296" xmlns="http://www.w3.org/2000/svg">
+    <g transform="matrix(1.4213 0 0 1.4213 -36.37 -21.406)">
+      <path d="m74.904 221.56c-0.18582-0.19116 0.02552-3.3476 0.46975-7.0142 1.2566-10.373 0.40827-23.378-1.8165-27.845-4.1258-8.2837-5.6136-15.201-7.847-36.48-2.5531-24.326-2.5558-24.337-12.097-51.954-7.2975-21.123-7.5036-22.275-5.1268-28.65 0.94139-2.525 4.0607-4.4793 7.1493-4.4793 4.6304 0 6.8049 2.7765 12.903 16.475 3.2707 7.3471 7.3576 16.456 9.082 20.243l3.1353 6.8846 0.76541-3.5378c1.7508-8.0921 7.0963-12.173 14.735-11.25 5.1089 0.61774 4.5571 0.93618 9.9236-5.7272 1.981-2.4598 3.3054-3.1698 6.6315-3.5555 2.2859-0.26508 5.121-0.23283 6.3002 0.07169 3.2842 0.84801 6.536 4.751 7.7864 9.3458 0.62101 2.2819 1.3748 3.8965 1.6752 3.5879 0.30034-0.30853 0.74682-15.914 0.9922-34.679 0.42267-32.324 0.54259-34.287 2.2806-37.331 0 0 0.91844-2.3872 1.8345-3.2128 1.8115-1.6326 4.2563-2.957 6.6949-2.9498 2.4064 0.0071 4.863 1.273 6.5889 2.9498 1.1042 1.0728 2.008 4.1591 2.008 4.1591 1.9523 4.0436 2.0332 5.3564 2.9132 47.274 0.49787 23.713 1.1164 44.561 1.3745 46.328 0.87291 5.9757 2.4478 10.061 4.0847 10.595 1.7786 0.58074 16.207-6.6556 15.528-7.7873-0.22012-0.36643 0.0919-0.66623 0.69333-0.66623 0.60145 0 2.7289-1.1134 4.7276-2.4741 5.307-3.6131 12.48-6.11 19.084-6.6424 4.9645-0.40027 6.3136-0.15897 9.3687 1.6756 3.8581 2.3169 6.2172 6.1713 5.3994 8.8221-0.28127 0.91174-2.2135 2.6166-4.2938 3.7886s-7.4322 5.9391-11.893 10.594c-4.9322 5.1462-10.527 9.919-14.276 12.179-7.7171 4.6513-15.614 11.984-23.861 22.155-7.6822 9.4756-9.4431 11.102-15.615 14.428-2.617 1.4099-5.2995 3.6404-5.9611 4.9566-0.73759 1.4674-1.448 7.5788-1.8364 15.798-0.55866 11.822-0.8233 13.456-2.2406 13.838-0.88394 0.2378-1.8312 0.20192-2.105-0.0798-0.27378-0.28166-0.49779-6.7894-0.49779-14.462v-13.95l-5.5766-0.484c-3.0671-0.26621-7.3834-1.1792-9.5918-2.029-2.2083-0.84975-4.5171-1.5835-5.1305-1.6306-0.61342-0.0471-0.93163-0.27454-0.70713-0.50551 0.2245-0.23096-0.47815-1.1592-1.5614-2.0628-3.4436-2.8724-0.94536-3.0045 5.7865-0.30608 10.577 4.2399 13.529 3.8283 25.343-3.5336 9.9919-6.2265 16.812-12.568 17.779-16.531 0.28922-1.1855 0.10206-3.3022-0.41589-4.7037-1.1771-3.1851-1.2308-8.8544-0.0839-8.8544 0.47182 0 1.3155 1.7556 1.8749 3.9013 0.55939 2.1457 1.2006 4.2108 1.425 4.589 0.22436 0.37825 3.2954-1.818 6.8246-4.8806 3.5292-3.0625 9.386-7.427 13.015-9.6988 4.1185-2.5782 9.4639-7.2126 14.223-12.331 4.1934-4.5102 8.3065-8.2004 9.1403-8.2004 0.83376 0 2.388-0.7265 3.4539-1.6144 2.4287-2.0233 1.7258-4.2521-1.8172-5.7619-4.7143-2.009-18.83 2.0141-24.791 7.0658-2.0105 1.7036-2.0385 1.8723-0.56297 3.3903 1.9257 1.9811 1.2526 2.8277-1.4923 1.8769-2.8372-0.9828-4.823-0.58034-10.551 2.1385-2.6991 1.2811-7.0154 3.3016-9.5918 4.4901s-4.7346 2.5791-4.7959 3.0904c-0.0613 0.51124-0.21191 1.3426-0.3346 1.8475-0.12268 0.50487-0.27325 1.3434-0.33459 1.8634-0.0613 0.51998-1.2922 0.0474-2.7352-1.0502-2.1881-1.6643-3.3362-1.8961-6.915-1.396-7.488 1.0463-12.259 3.121-13.95 6.0661-1.6963 2.9542-6.6386 4.9336-10.092 4.0418-3.004-0.77568-8.0112 1.2638-10.463 4.2617-1.1916 1.4569-2.6859 4.7339-3.3208 7.2823-0.65971 2.6482-1.6627 4.6334-2.341 4.6334-1.2146 0-1.1577-0.63006 0.77309-8.5575 0.55247-2.2684 2.3121-4.9456 4.9776-7.5731 3.94-3.8838 4.3859-5.5004 0.73336-2.6581-2.5494 1.9838-8.7309 1.7116-11.47-0.50511-1.5512-1.2553-3.6447-5.4739-6.4916-13.081-7.8955-21.097-7.5725-20.667-13.229-17.631-3.149 1.6895-4.4491 0.78245-1.6518-1.1523 2.6242-1.8151 2.5351-3.2273-0.63978-10.132-2.7335-5.9444-4.9171-7.8522-6.4374-5.6245-0.59928 0.8781-0.65932 0.41476-0.19086-1.4726 1.2275-4.9455 0.72312-8.0353-1.6389-10.04-1.4752-1.2519-2.4303-3.2411-2.8187-5.8706-0.68261-4.6211-4.5153-12.87-6.309-13.578-2.145-0.8468-4.8438 1.3985-5.391 4.4852-0.53974 3.0442 1.8795 10.771 10.271 32.805 5.1947 13.64 5.1496 13.535 6.0736 14.123 0.47647 0.30295 0.63883 4.0928 0.37741 8.8097-0.25233 4.5529-0.30271 8.2774-0.11195 8.2767 0.19075-6.9e-4 3.7322-0.61418 7.8699-1.3633 4.1377-0.74913 7.7179-1.1616 7.956-0.91669 0.60964 0.62719 0.39288 0.75173-6.352 3.6493-7.2396 3.1102-7.3755 3.453-6.5533 16.536 1.0063 16.012 4.4325 25.334 12.144 33.042l3.6739 3.6718-5.008 0.93068-1.4858 12.845c-0.8172 7.0647-1.5199 13.107-1.5615 13.426-0.08887 0.68308-3.3854 1.6742-3.8792 1.1662zm49.714-74.678c1.8294-2.0796 1.9233-8.1504 0.48358-31.257-1.0845-17.406-2.1058-22.356-5.1148-24.791-4.7953-3.8806-11.5-1.0648-14.092 5.9184-1.5534 4.1848-1.54 4.3949 1.2224 19.125 1.5362 8.1916 3.2823 17.785 3.8801 21.319 0.81703 4.83 1.6896 7.0525 3.5137 8.95 2.9747 3.0943 7.7278 3.44 10.107 0.73499zm-2.1897-4.8562c0-4.2063-0.0388-4.2625-3.1415-4.5438-3.8751-0.35143-4.3798-1.6834-1.0681-2.8191 1.8822-0.64548 3.0401-0.47523 4.7898 0.70421 2.0211 1.3624 2.2478 2.0136 1.8263 5.2462-0.73388 5.629-2.4065 6.6108-2.4065 1.4125zm-6.2458-22.742c1.4738-0.96498 6.8848-1.2558 5.9764-0.32123-0.31178 0.32075-2.0851 0.72141-3.9408 0.89035-2.3987 0.21839-2.9871 0.0539-2.0356-0.56912zm-8.332 25.159c1.3572-2.1309 1.4088-3.2109 0.4719-9.8679-1.5304-10.874-7.0715-32.764-8.7875-34.715-3.0438-3.4602-10.775-1.1886-12.764 3.7501-2.158 5.3598-1.2624 11.815 3.5849 25.839 5.9881 17.325 6.0288 17.386 11.55 17.386 3.7785 0 4.6414-0.34736 5.9441-2.3928zm-2.4132-4.4918c-1.171-2.361-1.8434-2.7384-4.7139-2.6459-3.692 0.11897-4.0063-1.2633-0.67871-2.985 4.0517-2.0963 7.6152 0.89231 7.0419 5.9058l-0.28346 2.479zm-6.8598-19.454c0.86814-0.23273 2.0727-0.21743 2.6768 0.034 0.60409 0.25145-0.10621 0.44187-1.5784 0.42316-1.4722-0.0187-1.9665-0.22444-1.0983-0.45717zm43.258 17.591c4.5394-0.86772 8.2534-1.8888 8.2534-2.269s-2.8106-1.4413-6.2458-2.3579c-3.4352-0.91664-6.2458-2.0353-6.2458-2.4859 0-0.46392 2.8616-0.54344 6.5971-0.18331l6.5971 0.636-1.3621-5.3769c-0.94743-3.7401-1.5297-12.333-1.9127-28.228-0.4882-20.259-0.72521-22.964-2.0894-23.847-1.1131-0.72072-2.2856-0.70636-4.2382 0.05192-2.0992 0.81516-2.6995 0.79565-2.6995-0.08775 0-0.62481 1.9072-1.7441 4.2382-2.4873l4.2382-1.3513 0.26455-7.2978c0.33429-9.2211-0.1686-11.26-2.9549-11.979-2.3486-0.60642-3.0653-2.2093-0.98792-2.2093 2.9706 0 3.4699-2.4079 2.4124-11.633-0.56124-4.8958-1.4586-9.4447-1.9941-10.108-1.7035-2.1117-5.9075-1.4582-7.5874 1.1793-1.3226 2.0767-1.6175 6.7242-2.273 35.818-0.5037 22.357-0.45287 34.009 0.15343 35.175 0.49866 0.95858 2.1099 2.0148 3.5806 2.3471 4.8397 1.0936 3.8459 2.2975-1.9373 2.347-5.378 0.046-5.5396 0.11245-3.6217 1.4889 2.8845 2.0701 2.5036 3.3106-0.78812 2.5668l-2.7957-0.63171 0.60332 14.098c0.73785 17.243 1.0815 19.356 3.0591 18.811 0.81556-0.22456 5.1969-1.1182 9.7362-1.986zm-49.96 59.654c5.4706-3.9421 11.543-5.0146 17.259-3.0484 5.5026 1.8927 4.7563 2.5594-2.845 2.5417-4.1282-0.01-8.9085 0.41905-10.623 0.95261-4.5946 1.43-6.1421 1.2479-3.7914-0.44595z" />
+      <path d="m78.915 31.501c-1.8054 0.50275-3.5949 1.6255-4.9013 3.0777-0.72939 0.81231-1.5317 2.0888-1.977 3.1462-0.98597 2.3319-1.3168 4.8578-1.0266 7.8386 0.50371 5.1633 3.0812 8.4592 8.7104 11.14 1.0386 0.49476 1.9904 0.90056 3.9908 1.6965 2.754 1.095 3.8264 1.5753 5.2735 2.3578 3.2651 1.7602 6.57 4.3695 7.9739 6.2921 0.16904 0.23242 0.34178 0.48663 0.38617 0.56455l0.08226 0.14344 0.04324-0.231c0.32739-1.8311 2.7241-5.6679 5.5787-8.9428 1.2507-1.4342 2.3739-2.5163 4.8364-4.6572 2.9368-2.5536 4.1044-3.7235 5.289-5.3127 1.7923-2.3997 2.5647-4.6567 2.5434-7.4302-5e-3 -0.69338-0.36744-2.9011-0.61764-3.7654-0.61029-2.0959-1.3487-3.6515-2.4581-5.156-0.78363-1.0601-1.6757-1.9279-2.7456-2.6684-1.8797-1.2992-3.8903-1.9098-6.2222-1.8907-0.59946 0.0075-2.3352 0.30186-2.9222 0.4954-3.3687 1.132-5.8962 3.5657-7.4543 7.1726-0.21002 0.48189-0.56352 1.4931-0.70263 2.0218l-0.11004 0.40301-0.58821-0.80506c-0.32142-0.44514-0.78883-1.0368-1.0383-1.3189-2.7542-3.106-6.2141-4.7048-9.8106-4.5212-0.47056 0.02246-1.6709 0.2212-2.1331 0.34984z" />
+    </g>
+  </svg>
+);
 
 export default Love;
