@@ -1,93 +1,91 @@
 import Topics from "@/components/common/Topics";
 
-export default function Terms() {
-  return (
-    <Topics>
-      <section>
-        <h2>Termos</h2>
-        <p>
-          Ao acessar ao site ALIS, concorda em cumprir estes termos de serviço, todas as leis e
-          regulamentos aplicáveis e concorda que é responsável pelo cumprimento de todas as leis
-          locais aplicáveis. Se você não concordar com algum desses termos, está proibido de usar ou
-          acessar este site. Os materiais contidos neste site estão em{" "}
-          <strong>domínio publico</strong>.
-        </p>
-      </section>
-      <section>
-        <h2>Uso de Licença</h2>
-        <p>
-          O ALIS está distribuído sobre a licença{" "}
-          <a href="https://unlicense.org/" rel="noopener noreferrer">
-            The Unlicense
-          </a>
-          . Isso significa que qualquer pessoa pode copiar, modificar, publicar, usar, vender ou
-          distribuir o código do ALIS, para qualquer finalidade que quiser.
-        </p>
-      </section>
-      <section>
-        <h2>Isenção de responsabilidade</h2>
-        <p>
-          Os materiais no site da ALIS são fornecidos &lsquo;como estão&rsquo;. ALIS não oferece
-          garantias, expressas ou implícitas, e, por este meio, isenta e nega todas as outras
-          garantias, incluindo, sem limitação, garantias implícitas ou condições de comercialização,
-          adequação a um fim específico ou não violação de propriedade intelectual ou outra violação
-          de direitos.
-        </p>
-        <p>
-          Além disso, o ALIS não garante ou faz qualquer representação relativa à precisão, aos
-          resultados prováveis ou à confiabilidade do uso dos materiais em seu site ou de outra
-          forma relacionado a esses materiais ou em sites vinculados a este site.
-        </p>
-      </section>
-      <section>
-        <h2>Limitações</h2>
-        <p>
-          Em nenhum caso o ALIS ou seus contribuidores serão responsáveis por quaisquer danos
-          (incluindo, sem limitação, danos por perda de dados ou lucro ou devido a interrupção dos
-          negócios) decorrentes do uso ou da incapacidade de usar os materiais em ALIS, mesmo que
-          ALIS ou um representante autorizado da ALIS tenha sido notificado oralmente ou por escrito
-          da possibilidade de tais danos.
-        </p>
-        <p>
-          Como algumas jurisdições não permitem limitações em garantias implícitas, ou limitações de
-          responsabilidade por danos consequentes ou incidentais, essas limitações podem não se
-          aplicar a você.
-        </p>
-      </section>
-      <section>
-        <h2>Precisão dos materiais</h2>
-        <p>
-          Os materiais exibidos no site da ALIS podem incluir erros técnicos, tipográficos ou
-          fotográficos. ALIS e seus contribuidores não garantem que qualquer material em seu site
-          seja preciso, completo ou atual. ALIS pode fazer alterações nos materiais contidos em seu
-          site a qualquer momento, sem aviso prévio. No entanto, ALIS não se compromete a atualizar
-          os materiais.
-        </p>
-      </section>
-      <section>
-        <h2>Links</h2>
-        <p>
-          O ALIS não analisou todos os sites vinculados ao seu site e não é responsável pelo
-          conteúdo de nenhum site vinculado. A inclusão de qualquer link não implica endosso por
-          ALIS do site. O uso de qualquer site vinculado é por conta e risco do usuário.
-        </p>
-      </section>
-      <section>
-        <h2>Modificações</h2>
-        <p>
-          O ALIS pode revisar estes termos de serviço do site a qualquer momento, sem aviso prévio.
-          Ao usar este site, você concorda em ficar vinculado à versão atual desses termos de
-          serviço.
-        </p>
-      </section>
-      <section>
-        <h2>Lei aplicável</h2>
-        <p>
-          Estes termos e condições são regidos e interpretados de acordo com as leis do ALIS e você
-          se submete irrevogavelmente à jurisdição exclusiva dos tribunais naquele estado ou
-          localidade.
-        </p>
-      </section>
-    </Topics>
-  );
-}
+const Terms = () => (
+  <Topics>
+    <section>
+      <h2>Termos</h2>
+      <p>
+        Ao acessar ao site ALIS, concorda em cumprir estes termos de serviço, todas as leis e
+        regulamentos aplicáveis e concorda que é responsável pelo cumprimento de todas as leis
+        locais aplicáveis. Se você não concordar com algum desses termos, está proibido de usar ou
+        acessar este site. Os materiais contidos neste site estão em{" "}
+        <strong>domínio publico</strong>.
+      </p>
+    </section>
+    <section>
+      <h2>Uso de Licença</h2>
+      <p>
+        O ALIS está distribuído sobre a licença{" "}
+        <a href="https://unlicense.org/" rel="noopener noreferrer">
+          The Unlicense
+        </a>
+        . Isso significa que qualquer pessoa pode copiar, modificar, publicar, usar, vender ou
+        distribuir o código do ALIS, para qualquer finalidade que quiser.
+      </p>
+    </section>
+    <section>
+      <h2>Isenção de responsabilidade</h2>
+      <p>
+        Os materiais no site da ALIS são fornecidos &lsquo;como estão&rsquo;. ALIS não oferece
+        garantias, expressas ou implícitas, e, por este meio, isenta e nega todas as outras
+        garantias, incluindo, sem limitação, garantias implícitas ou condições de comercialização,
+        adequação a um fim específico ou não violação de propriedade intelectual ou outra violação
+        de direitos.
+      </p>
+      <p>
+        Além disso, o ALIS não garante ou faz qualquer representação relativa à precisão, aos
+        resultados prováveis ou à confiabilidade do uso dos materiais em seu site ou de outra forma
+        relacionado a esses materiais ou em sites vinculados a este site.
+      </p>
+    </section>
+    <section>
+      <h2>Limitações</h2>
+      <p>
+        Em nenhum caso o ALIS ou seus contribuidores serão responsáveis por quaisquer danos
+        (incluindo, sem limitação, danos por perda de dados ou lucro ou devido a interrupção dos
+        negócios) decorrentes do uso ou da incapacidade de usar os materiais em ALIS, mesmo que ALIS
+        ou um representante autorizado da ALIS tenha sido notificado oralmente ou por escrito da
+        possibilidade de tais danos.
+      </p>
+      <p>
+        Como algumas jurisdições não permitem limitações em garantias implícitas, ou limitações de
+        responsabilidade por danos consequentes ou incidentais, essas limitações podem não se
+        aplicar a você.
+      </p>
+    </section>
+    <section>
+      <h2>Precisão dos materiais</h2>
+      <p>
+        Os materiais exibidos no site da ALIS podem incluir erros técnicos, tipográficos ou
+        fotográficos. ALIS e seus contribuidores não garantem que qualquer material em seu site seja
+        preciso, completo ou atual. ALIS pode fazer alterações nos materiais contidos em seu site a
+        qualquer momento, sem aviso prévio. No entanto, ALIS não se compromete a atualizar os
+        materiais.
+      </p>
+    </section>
+    <section>
+      <h2>Links</h2>
+      <p>
+        O ALIS não analisou todos os sites vinculados ao seu site e não é responsável pelo conteúdo
+        de nenhum site vinculado. A inclusão de qualquer link não implica endosso por ALIS do site.
+        O uso de qualquer site vinculado é por conta e risco do usuário.
+      </p>
+    </section>
+    <section>
+      <h2>Modificações</h2>
+      <p>
+        O ALIS pode revisar estes termos de serviço do site a qualquer momento, sem aviso prévio. Ao
+        usar este site, você concorda em ficar vinculado à versão atual desses termos de serviço.
+      </p>
+    </section>
+    <section>
+      <h2>Lei aplicável</h2>
+      <p>
+        Estes termos e condições são regidos e interpretados de acordo com as leis do ALIS e você se
+        submete irrevogavelmente à jurisdição exclusiva dos tribunais naquele estado ou localidade.
+      </p>
+    </section>
+  </Topics>
+);
+
+export default Terms;
