@@ -8,24 +8,24 @@ interface Props {
   };
 }
 
-export default function UserCard({ user }: Props) {
-  return (
-    <article className={styles.card}>
-      <div>
-        <h1>{user.name}</h1>
-        <p>{user?.email} </p>
-        <p>
-          Edite suas informações pelo seu
-          <a href="https://myaccount.google.com/"> Perfil do Google</a>
-        </p>
-      </div>
-      <div>
-        <img
-          src={user?.image !== null ? user?.image : undefined}
-          data-cy="profileImage"
-          referrerPolicy="no-referrer"
-        />
-      </div>
-    </article>
-  );
-}
+const UserCard = ({ user }: Props) => (
+  <article className={styles.card}>
+    <div>
+      <h1>{user.name}</h1>
+      <p>{user?.email} </p>
+      <p>
+        Edite suas informações pelo seu
+        <a href="https://myaccount.google.com/"> Perfil do Google</a>
+      </p>
+    </div>
+    <div>
+      <img
+        data-cy="profileImage"
+        referrerPolicy="no-referrer"
+        src={user?.image !== null ? user?.image : undefined}
+      />
+    </div>
+  </article>
+);
+
+export default UserCard;
