@@ -5,27 +5,25 @@ import Header from "@/components/common/Header";
 import Title from "@/components/common/Title";
 import Questions from "@/components/help/Questions";
 
-const Privacy = () => {
-  return (
-    <>
-      <Head>
-        <title>Perguntas frequentes</title>
-      </Head>
+const Privacy = () => (
+  <>
+    <Head>
+      <title>Perguntas frequentes</title>
+    </Head>
 
-      <Header />
+    <Header />
 
-      <main>
-        <Title small>
-          <h1>
-            Perguntas <span>frequentes</span>
-          </h1>
-        </Title>
+    <main>
+      <Title small>
+        <h1>
+          Perguntas <span>frequentes</span>
+        </h1>
+      </Title>
 
-        <Questions />
-      </main>
-      <Footer />
-    </>
-  );
-};
+      <Questions />
+    </main>
+    <Footer />
+  </>
+);
 
 export default Privacy;
