@@ -13,30 +13,28 @@ interface Props {
   place: Place;
 }
 
-const PlacePage = ({ place }: Props) => {
-  return (
-    <>
-      <Head>
-        <title>Editar {place.name}</title>
-      </Head>
+const PlacePage = ({ place }: Props) => (
+  <>
+    <Head>
+      <title>Editar {place.name}</title>
+    </Head>
 
-      <Header />
+    <Header />
 
-      <main>
-        <Title>
-          <h1>
-            Atualizar <span>dados</span>
-          </h1>
-          <p>Modifique os dados do Local por aqui</p>
-        </Title>
+    <main>
+      <Title>
+        <h1>
+          Atualizar <span>dados</span>
+        </h1>
+        <p>Modifique os dados do Local por aqui</p>
+      </Title>
 
-        <FormChange place={place} />
-      </main>
+      <FormChange place={place} />
+    </main>
 
-      <Footer />
-    </>
-  );
-};
+    <Footer />
+  </>
+);
 
 export default PlacePage;
 
